+++ server/api/message/message.integration.js
@@ -21,9 +21,7 @@ describe('Message API:', function() {
         .expect(200)
         .expect('Content-Type', /html/)
         .end((err, res) => {
-          if(err) {
-            return done(err);
-          }
+          if(err) return done(err);
           response = res.text;
           done();
         });